Improvements

+++ src/Board.h
@@ -26,7 +26,7 @@ private:
 	sf::RenderWindow& m_window;
 
 	size_t gridToIndex(sf::Vector2i pos) { return (pos.y * Settings::cellAmount) + pos.x; }
-	sf::Vector2i indexToGridPos(int index) { return {index % Settings::cellAmount, index / Settings::cellAmount}; }
+	sf::Vector2i indexToGridPos(size_t index) { return {static_cast<int>(index % Settings::cellAmount), static_cast<int>(index / Settings::cellAmount)}; }
 
 	void nextGen();
